NF: move MutableTime to MockTime
Used only once. No need to restrict mocking
