+++ AnkiDroid/src/test/java/com/ichi2/libanki/sched/SchedTest.java
@@ -32,7 +32,6 @@ import com.ichi2.libanki.Model;
 import com.ichi2.libanki.Models;
 import com.ichi2.libanki.Note;
 import com.ichi2.testutils.MockTime;
-import com.ichi2.testutils.MutableTime;
 import com.ichi2.utils.JSONArray;
 import com.ichi2.utils.JSONObject;
 
@@ -186,10 +185,10 @@ public class SchedTest extends RobolectricTest {
 
     @Test
     public void testRevLogValues() {
-        MutableTime time = new MutableTime(MockTime.timeStamp(2020, 8, 4, 11, 22, 19, 123), 10);
-        Collection col =  CollectionHelper.getInstance().getCol(getTargetContext(), time);
+        Collection col = getCol();
         addNoteUsingBasicModel("Hello", "World");
 
+        MockTime time = (MockTime) col.getTime();
         AbstractSched sched = col.getSched();
         Card c = sched.getCard();
         time.setFrozen(true);
